% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_metadata.r
\name{mv_add_metadata}
\alias{mv_add_metadata}
\title{Add Metadata to Data Object}
\usage{
mv_add_metadata(data, ..., .replace = TRUE)
}
\arguments{
\item{data}{Data object (typically data frame)}

\item{...}{Named metadata items to add}

\item{.replace}{Logical, whether to replace existing metadata}
}
\value{
The data object with metadata attributes added
}
\description{
Add Metadata to Data Object
}
