% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_messaging.r
\name{mv_alert}
\alias{mv_alert}
\title{Alert message}
\usage{
mv_alert(
  message,
  type = c("info", "warning", "error"),
  ...,
  .envir = parent.frame()
)
}
\arguments{
\item{message}{Main alert message (scalar string)}

\item{type}{One of "info", "warning", "error"}

\item{...}{Additional bullet components using cli syntax}

\item{.envir}{Environment for string interpolation}
}
\description{
Alert message
}
