% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_config.r
\name{mv_get_config}
\alias{mv_get_config}
\title{Get macroverse Configuration Options}
\usage{
mv_get_config(option = NULL, default = NULL)
}
\arguments{
\item{option}{Character, specific option to retrieve. If NULL, returns all options.}

\item{default}{Default value if option is not set}
}
\value{
The option value, or a list of all options if option is NULL
}
\description{
Get macroverse Configuration Options
}
