% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_metadata.r
\name{mv_get_metadata}
\alias{mv_get_metadata}
\title{Get Metadata from Data Object}
\usage{
mv_get_metadata(data, key = NULL, .remove_prefix = TRUE)
}
\arguments{
\item{data}{Data object with metadata}

\item{key}{Optional specific metadata key to retrieve}

\item{.remove_prefix}{Logical, remove the prefix from returned names}
}
\value{
If key is provided, returns the specific metadata value.
Otherwise returns a list of all metadata.
}
\description{
Get Metadata from Data Object
}
