% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_metadata.r
\name{mv_has_metadata}
\alias{mv_has_metadata}
\title{Check if Data Has Metadata}
\usage{
mv_has_metadata(data, key = NULL)
}
\arguments{
\item{data}{Data object to check}

\item{key}{Optional specific metadata key to check for}
}
\value{
Logical indicating whether metadata exists
}
\description{
Check if Data Has Metadata
}
