% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_messaging.r
\name{mv_inform}
\alias{mv_inform}
\title{Information message}
\usage{
mv_inform(message, ..., .envir = parent.frame())
}
\arguments{
\item{message}{Main information message}

\item{...}{Additional message components using cli syntax}

\item{.envir}{Environment for string interpolation}
}
\description{
Information message
}
