% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_metadata.r
\name{mv_remove_metadata}
\alias{mv_remove_metadata}
\title{Remove Metadata from Data Object}
\usage{
mv_remove_metadata(data, keys = NULL)
}
\arguments{
\item{data}{Data object with metadata}

\item{keys}{Character vector of metadata keys to remove. If NULL, removes all.}
}
\value{
The data object with metadata removed
}
\description{
Remove Metadata from Data Object
}
