% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_config.r
\name{mv_set_config}
\alias{mv_set_config}
\title{Set macroverse Configuration Options}
\usage{
mv_set_config(..., .reset = FALSE)
}
\arguments{
\item{...}{Named arguments for options to set}

\item{.reset}{Logical, reset all options to defaults}
}
\value{
Invisibly returns the previous values of changed options
}
\description{
Set macroverse Configuration Options
}
