% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_messaging.r
\name{mv_stop}
\alias{mv_stop}
\title{Stop with formatted error message}
\usage{
mv_stop(message, ..., .envir = parent.frame())
}
\arguments{
\item{message}{Main error message}

\item{...}{Additional message components using cli syntax}

\item{.envir}{Environment for string interpolation}
}
\description{
Stop with formatted error message
}
