% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_utils.r
\name{mv_system_info}
\alias{mv_system_info}
\title{Get System Information}
\usage{
mv_system_info()
}
\value{
List with system information
}
\description{
Returns useful system information for debugging and support.
}
