% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_messaging.r
\name{mv_warn}
\alias{mv_warn}
\title{Warning with formatted message}
\usage{
mv_warn(message, ..., .envir = parent.frame())
}
\arguments{
\item{message}{Main warning message}

\item{...}{Additional message components using cli syntax}

\item{.envir}{Environment for string interpolation}
}
\description{
Warning with formatted message
}
