% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_messaging.r
\name{mv_alert}
\alias{mv_alert}
\title{Show alert message}
\usage{
mv_alert(message, type = "info", .envir = parent.frame())
}
\arguments{
\item{message}{Alert message}

\item{type}{Type of alert: "info", "success", "warning", or "danger"}

\item{.envir}{Environment for string interpolation}
}
\description{
Show alert message
}
